+++ odoo/local-src/bso_mailchimp/models/mailchimp_list_segment.py
@@ -3,7 +3,6 @@ import logging
 from datetime import datetime
 
 from odoo import fields, models, api
-
 from . import mailchimp_client
 
 _logger = logging.getLogger(__name__)
@@ -90,30 +89,27 @@ class MailchimpListSegment(models.Model):
 
     @api.multi
     def write(self, values):
-        record = super(MailchimpListSegment, self).write(values)
-        if 'mailchimp_ref' in values:
-            return record  # Values are coming from Mailchimp -> Don't update
-
         client = self.env['mailchimp.client'].get_client()
         for rec in self:
+            if 'mailchimp_ref' in values:
+                continue  # Values are coming from Mailchimp
+                # -> Don't update
             if 'lead_ids' in values:
+                edited_segment = self.new(values)
+                remaining_leads = edited_segment.lead_ids
                 saved_leads = rec.lead_ids
-                remaining_lead_ids = values['lead_ids'][0][2]
-                remaining_leads = self.env["crm.lead"].browse(
-                    remaining_lead_ids)
                 unlinked_leads = saved_leads - remaining_leads
-                members_to_remove = rec.get_members(unlinked_leads)
                 added_leads = remaining_leads - saved_leads
+                members_to_remove = rec.get_members(unlinked_leads)
                 members_to_add = rec.get_members(added_leads)
                 rec._update_members(client, members_to_add, members_to_remove)
             if 'name' in values:
-                rec._update_name(client)
+                rec._update_name(client, values)
+        return super(MailchimpListSegment, self).write(values)
 
-        return record
-
-    def _update_name(self, client):
+    def _update_name(self, client, values):
         data = {
-            "name": self.name,
+            "name": values.get('name', self.name),
         }
         return client.lists.segments.update(
             self.list_id.mailchimp_ref, self.mailchimp_ref, data)